refactor(backend): use entities to manage amadeus response
