+++ backend/src/main/java/encora/breakableII/backend/entities/FlightAmadeus.java
@@ -1,6 +1,7 @@
-package encora.breakableII.backend.models;
+package encora.breakableII.backend.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import encora.breakableII.backend.models.Price;
 
 import java.util.List;
