+++ backend/src/main/java/encora/breakableII/backend/entities/Location.java
@@ -1,4 +1,4 @@
-package encora.breakableII.backend.models;
+package encora.breakableII.backend.entities;
 
 public class Location {
     private String cityCode;
