+++ backend/src/main/java/encora/breakableII/backend/entities/TokenResponse.java
@@ -1,4 +1,4 @@
-package encora.breakableII.backend.models;
+package encora.breakableII.backend.entities;
 
 public class TokenResponse {
     private String access_token;
